reorganizing test files
