+++ contracts/marketplace/src/testing/helpers/nft_functions.rs
@@ -1,5 +1,9 @@
+use std::collections::HashSet;
+use std::iter::FromIterator;
+
 use cosmwasm_std::coins;
 use cosmwasm_std::{Addr, Empty};
+use cw721::TokensResponse;
 use cw_multi_test::Executor;
 use sg721::ExecuteMsg as Sg721ExecuteMsg;
 use sg721_base::msg::CollectionInfoResponse;
@@ -78,4 +82,25 @@ pub fn burn(router: &mut StargazeApp, creator: &Addr, collection: &Addr, token_i
     };
     let res = router.execute_contract(creator.clone(), collection.clone(), &transfer_msg, &[]);
     assert!(res.is_ok());
+}
+
+pub fn get_next_token_id_and_map(
+    router: &mut StargazeApp,
+    incoming_hash: &HashSet<String>,
+    collection: Addr,
+) -> (HashSet<std::string::String>, u32) {
+    let query_msg = sg721_base::msg::QueryMsg::AllTokens {
+        start_after: None,
+        limit: None,
+    };
+    let res: TokensResponse = router
+        .wrap()
+        .query_wasm_smart(collection, &query_msg)
+        .unwrap();
+    let tokens_hash: HashSet<String> = HashSet::from_iter(res.tokens.iter().cloned());
+    let difference = tokens_hash.difference(incoming_hash);
+    let nft_hash = tokens_hash.clone();
+    let token_id: Option<&String> = difference.into_iter().next();
+    let token_id_unwrapped = token_id.unwrap().parse::<u32>().unwrap();
+    (nft_hash, token_id_unwrapped)
 }
